Somewhat large reorganisation, trying to get a focused segmentation interface and clear separation with its first implementation (impl0) (e.g. contours is an implementation detail and as such should not be part of the interface).
Decrease the amount of generated code : trying to keep it down to what would be boring or impossible to generate by hand, the rest -e.g. logic- should be in hand-written code.

Some CMake modernisation.

And finally started to implement in impl0 zones (group of motif positions of a given pad size) to speed up hasPadByPosition to be more in line with aliroot perf (mostly for St1 B, as the other ones are already on par or better than aliroot)